port: make client inventory links native-width

+++ src/legacy/vardefs.c
@@ -640,6 +640,6 @@ uint32_t dword_587000_122956 = 0x1;
 uint32_t dword_5d4594_2491716 = 0;
 uint32_t dword_587000_180476 = 0x0;
 uint32_t dword_5d4594_1046852 = 0;
-uint32_t array_5D4594_1049872[9];
+uintptr_t array_5D4594_1049872[9];
 uintptr_t dword_8531A0_2576 = 0;
 uint32_t dword_8531A0_2572 = 0;
